Installed and configured dependencies for testing
This includes babel. Babel transpiles javascript to an older version during runtime, so newer syntax can be used without any hassle. This enables ES6 by default including ES6 modules which weren't working that well earlier. This is also necessary to run jest tests with ES6 modules.

Also installed jest and supertest. Jest is for running the tests and supertest is for running the api for the integration tests.

eslint-plugin-jest is also required for eslint to recognize jest functions.