added tests for the delete, moveUp and moveDown buttons
the tests use an edited version of the mockMinutesContextState just so that there are 3 segments. The rendering also isnt done in beforeAlls except for the delete tests. That is because the up and down button describes needed to have different segment indexes in the tests to test all the functionalities